[Silver II] Title: A → B, Time: 16 ms, Memory: 5116 KB -BaekjoonHub

+++ 백준/Silver/16953. A → B/A → B.cc
@@ -17,7 +17,7 @@ bool isRange(long long x){
     return 1 <= x && x <= 1e9; 
 }
 
-long long bfs(){
+int bfs(){
     queue<p> q;
     q.push({a, 0});
     visited[a] = 1;
